requirements.txt
openai-whisper>=20230314
tqdm
ffmpeg-python